Termino el Historial de Cliente

+++ src/Objetos/ClienteHistorial.cs
@@ -9,7 +9,7 @@ namespace PalcoNet.Objetos
     public class ClienteHistorial {
         public int compra_id { get; set; }
         public DateTime compra_fecha { get; set; }
-        public int item_monto { get; set; }
+        public float item_monto { get; set; }
         public string fact_pago_desc { get; set; }
     }
 }
